Логин/регистрация в edx перенесены в courses
