+++ package/courses/delegate.go
@@ -1,6 +1,9 @@
 package courses
 
-import "github.com/skinnykaen/robbo_student_personal_account.git/package/models"
+import (
+	"github.com/skinnykaen/robbo_student_personal_account.git/package/edxApi"
+	"github.com/skinnykaen/robbo_student_personal_account.git/package/models"
+)
 
 type Delegate interface {
 	CreateCourse(course *models.CourseHTTP, courseId string) (id string, err error)
@@ -12,4 +15,6 @@ type Delegate interface {
 	GetEnrollments(username string) (respBody []byte, err error)
 	PostUnenroll(postUnenrollHTTP *models.PostEnrollmentHTTP) (err error)
 	PostEnrollment(postEnrollmentHTTP *models.PostEnrollmentHTTP) (err error)
+	Registration(userForm *edxApi.RegistrationForm) (err error)
+	Login(email, password string) (err error)
 }
